requirements.txt
# Core dependencies
pyspark==3.5.0
kafka-python==2.0.2
redis==5.0.1
psycopg2-binary==2.9.9
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
joblib==1.3.2

# Delta Lake
delta-spark==3.0.0

# S3/MinIO support
s3fs==2023.12.2
boto3==1.34.0

# Schema validation
jsonschema==4.20.0

# Monitoring
prometheus-client==0.19.0

# Web framework
streamlit==1.29.0
plotly==5.18.0

# ML/Data
sqlalchemy==2.0.23

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
faker==20.1.0
